Migrate thankwords command
Adjustements needed after merge of #357
