+++ src/main/java/de/chojo/repbot/ReputationBot.java
@@ -27,8 +27,8 @@ import de.chojo.repbot.commands.reputation.Reputation;
 import de.chojo.repbot.commands.Roles;
 import de.chojo.repbot.commands.scan.Scan;
 import de.chojo.repbot.commands.Setup;
-import de.chojo.repbot.commands.Thankwords;
 import de.chojo.repbot.commands.top.Top;
+import de.chojo.repbot.commands.thankwords.Thankwords;
 import de.chojo.repbot.config.Configuration;
 import de.chojo.repbot.dao.access.Cleanup;
 import de.chojo.repbot.dao.provider.Guilds;
