+++ src/main/java/de/chojo/repbot/commands/thankwords/handler/Remove.java
@@ -1,0 +1,50 @@
+package de.chojo.repbot.commands.thankwords.handler;
+
+import de.chojo.jdautil.interactions.slash.structure.handler.SlashHandler;
+import de.chojo.jdautil.localization.util.Format;
+import de.chojo.jdautil.localization.util.Replacement;
+import de.chojo.jdautil.util.Completion;
+import de.chojo.jdautil.wrapper.EventContext;
+import de.chojo.repbot.dao.provider.Guilds;
+import net.dv8tion.jda.api.events.interaction.command.CommandAutoCompleteInteractionEvent;
+import net.dv8tion.jda.api.events.interaction.command.SlashCommandInteractionEvent;
+
+import java.util.regex.Pattern;
+import java.util.regex.PatternSyntaxException;
+
+public class Remove implements SlashHandler {
+    private final Guilds guilds;
+
+    public Remove(Guilds guilds) {
+
+        this.guilds = guilds;
+    }
+
+    @Override
+    public void onSlashCommand(SlashCommandInteractionEvent event, EventContext context) {
+        var pattern = event.getOption("pattern").getAsString();
+        try {
+            Pattern.compile(pattern);
+        } catch (PatternSyntaxException e) {
+            event.reply(context.localize("error.invalidRegex"))
+                    .setEphemeral(true)
+                    .queue();
+            return;
+        }
+        if (guilds.guild(event.getGuild()).settings().thanking().thankwords().remove(pattern)) {
+            event.reply(context.localize("command.thankwords.remove.message.removed",
+                    Replacement.create("PATTERN", pattern, Format.CODE))).queue();
+            return;
+        }
+        event.reply(context.localize("command.thankwords.remove.message.patternnotnound"))
+                .setEphemeral(true)
+                .queue();
+    }
+
+    @Override
+    public void onAutoComplete(CommandAutoCompleteInteractionEvent event, EventContext context) {
+        var option = event.getFocusedOption();
+        var thankwords = guilds.guild(event.getGuild()).settings().thanking().thankwords().words();
+        event.replyChoices(Completion.complete(option.getValue(), thankwords)).queue();
+    }
+}
