+++ src/main/resources/locale.properties
@@ -275,34 +275,34 @@ command.setup.dialog.setupComplete=
 command.setup.dialog.starting=
 command.setup.dialog.welcome=
 command.thankwords.description=
-command.thankwords.error.patternNotFound=
-command.thankwords.sub.add=
-command.thankwords.sub.add.added=
-command.thankwords.sub.add.arg.pattern=
-command.thankwords.sub.check=
-command.thankwords.sub.check.arg.message=
-command.thankwords.sub.check.confidence=
-command.thankwords.sub.check.match.answer=
-command.thankwords.sub.check.match.fuzzy=
-command.thankwords.sub.check.match.mention=
-command.thankwords.sub.check.match.noMatch=
+command.thankwords.remove.message.patternnotnound=
+command.thankwords.add.description=
+command.thankwords.add.message.added=
+command.thankwords.add.pattern.description=
+command.thankwords.check.description=
+command.thankwords.check.message.description=
+command.thankwords.check.message.confidence=
+command.thankwords.check.message.answer=
+command.thankwords.check.message.fuzzy=
+command.thankwords.check.message.mention=
+command.thankwords.check.message.noMatch=
 command.thankwords.sub.check.notConfident=
-command.thankwords.sub.check.reference=
-command.thankwords.sub.check.result=
-command.thankwords.sub.list=
-command.thankwords.sub.list.list=
-command.thankwords.sub.loadDefault=
-command.thankwords.sub.loadDefault.added=
-command.thankwords.sub.loadDefault.arg.language=
-command.thankwords.sub.loadDefault.available=
-command.thankwords.sub.remove=
-command.thankwords.sub.remove.arg.pattern=
-command.thankwords.sub.remove.removed=
 command.top.message.empty=
 command.top.message.monthTitle=
 command.top.message.total=
 command.top.message.weekTitle=
 command.top.mode.description=
+command.thankwords.check.message.reference=
+command.thankwords.check.message.result=
+command.thankwords.list.description=
+command.thankwords.list.message.list=
+command.thankwords.loaddefault.description=
+command.thankwords.loaddefault.message.added=
+command.thankwords.loaddefault.language.description=
+command.thankwords.loaddefault.message.available=
+command.thankwords.remove.description=
+command.thankwords.remove.pattern.description=
+command.thankwords.remove.message.removed=
 conversation.cancel=
 conversation.canceled=
 conversation.exit=
