+++ src/main/resources/locale_de_DE.properties
@@ -269,34 +269,34 @@ command.setup.dialog.setupComplete=Einrichtung abgeschlossen!
 command.setup.dialog.starting=Einrichtung wird gestartet.
 command.setup.dialog.welcome=Willkommen bei der Konfiguration des Reputations-Bots.
 command.thankwords.description=Konfiguriert die Dankwörter.
-command.thankwords.error.patternNotFound=Muster existiert nicht.
-command.thankwords.sub.add=Füge ein Muster hinzu. Diese werden als Regex interpretiert. Keine Leerzeichen!
-command.thankwords.sub.add.added=Muster %REGEX% hinzugefügt.
-command.thankwords.sub.add.arg.pattern=Das Regex-Muster, das hinzugefügt werden soll.
-command.thankwords.sub.check=Prüft, ob eine Meldung Reputation geben würde.
-command.thankwords.sub.check.arg.message=Die Nachricht, die mit den aktuellen Dankwörtern im aktuellen Kanal überprüft werden soll.
-command.thankwords.sub.check.confidence=Zuversichtlichkeit: %SCORE%
-command.thankwords.sub.check.match.answer=Antwortübereinstimmung
-command.thankwords.sub.check.match.fuzzy=Grobe Übereinstimmung
-command.thankwords.sub.check.match.mention=Erwähnung gefunden
-command.thankwords.sub.check.match.noMatch=Kein Treffer
+command.thankwords.remove.message.patternnotnound=Muster existiert nicht.
+command.thankwords.add.description=Füge ein Muster hinzu. Diese werden als Regex interpretiert. Keine Leerzeichen!
+command.thankwords.add.message.added=Muster %REGEX% hinzugefügt.
+command.thankwords.add.pattern.description=Das Regex-Muster, das hinzugefügt werden soll.
+command.thankwords.check.description=Prüft, ob eine Meldung Reputation geben würde.
+command.thankwords.check.message.description=Die Nachricht, die mit den aktuellen Dankwörtern im aktuellen Kanal überprüft werden soll.
+command.thankwords.check.message.confidence=Zuversichtlichkeit: %SCORE%
+command.thankwords.check.message.answer=Antwortübereinstimmung
+command.thankwords.check.message.fuzzy=Grobe Übereinstimmung
+command.thankwords.check.message.mention=Erwähnung gefunden
+command.thankwords.check.message.noMatch=Kein Treffer
 command.thankwords.sub.check.notConfident=Zuversichtlichkeit ist zu niedrig. Es würde keine Reputation verteilt werden.
-command.thankwords.sub.check.reference=Referenznachricht: [Klick](%URL%)
-command.thankwords.sub.check.result=Spender: %DONATOR%\nEmpfänger: %RECEIVER%
-command.thankwords.sub.list=Listet alle Muster auf
-command.thankwords.sub.list.list=Folgende Muster sind aktiv:
-command.thankwords.sub.loadDefault=Lade Standard Dankeswörter für eine Sprache.
-command.thankwords.sub.loadDefault.added=Folgende Dankeswörter würden hinzugefügt:
-command.thankwords.sub.loadDefault.arg.language=Die Sprache, für die Dankwörter geladen werden sollen.
-command.thankwords.sub.loadDefault.available=Verfügbare Sprachen:
-command.thankwords.sub.remove=Löscht ein Muster
-command.thankwords.sub.remove.arg.pattern=Das Regex-Muster, das entfernt werden soll.
-command.thankwords.sub.remove.removed=Muster %PATTERN% wurde entfernt.
 command.top.message.empty=Es sind keine Einträge vorhanden
 command.top.message.monthTitle=Rangliste von %GUILD% für die letzten 30 Tage
 command.top.message.total=Rangliste für %GUILD%
 command.top.message.weekTitle=Rangliste von %GUILD% für die letzten 7 Tage.
 command.top.mode.description=$command.reputation.description.arg.mode$
+command.thankwords.check.message.reference=Referenznachricht: [Klick](%URL%)
+command.thankwords.check.message.result=Spender: %DONATOR%\nEmpfänger: %RECEIVER%
+command.thankwords.list.description=Listet alle Muster auf
+command.thankwords.list.message.list=Folgende Muster sind aktiv:
+command.thankwords.loaddefault.description=Lade Standard Dankeswörter für eine Sprache.
+command.thankwords.loaddefault.message.added=Folgende Dankeswörter würden hinzugefügt:
+command.thankwords.loaddefault.language.description=Die Sprache, für die Dankwörter geladen werden sollen.
+command.thankwords.loaddefault.message.available=Verfügbare Sprachen:
+command.thankwords.remove.description=Löscht ein Muster
+command.thankwords.remove.pattern.description=Das Regex-Muster, das entfernt werden soll.
+command.thankwords.remove.message.removed=Muster %PATTERN% wurde entfernt.
 conversation.cancel=Abbrechen
 conversation.canceled=Konversation abgebrochen.
 conversation.exit=Beenden
