+++ src/main/resources/locale_en_US.properties
@@ -275,33 +275,33 @@ command.setup.dialog.setupComplete=Setup complete!
 command.setup.dialog.starting=Starting setup.
 command.setup.dialog.welcome=Welcome to the configuration of the reputation bot.
 command.thankwords.description=Configure Thankwords
-command.thankwords.error.patternNotFound=Pattern does not exist.
-command.thankwords.sub.add=Add another pattern. Pattern are interpreted as Regex. A pattern can not contain spaces.
-command.thankwords.sub.add.added=Added pattern %REGEX%.
-command.thankwords.sub.add.arg.pattern=The regex pattern to add.
-command.thankwords.sub.check=Check if a message would give reputation.
-command.thankwords.sub.check.arg.message=The message to check with the current thankwords in the current channel.
-command.thankwords.sub.check.confidence=Confidence Score: %SCORE%
-command.thankwords.sub.check.match.answer=Answer Match
-command.thankwords.sub.check.match.fuzzy=Fuzzy Match
-command.thankwords.sub.check.match.mention=Mention Match
-command.thankwords.sub.check.match.noMatch=No Match
+command.thankwords.remove.message.patternnotnound=Pattern does not exist.
+command.thankwords.add.description=Add another pattern. Pattern are interpreted as Regex. A pattern can not contain spaces.
+command.thankwords.add.message.added=Added pattern %REGEX%.
+command.thankwords.add.pattern.description=The regex pattern to add.
+command.thankwords.check.description=Check if a message would give reputation.
+command.thankwords.check.message.description=The message to check with the current thankwords in the current channel.
+command.thankwords.check.message.confidence=Confidence Score: %SCORE%
+command.thankwords.check.message.answer=Answer Match
+command.thankwords.check.message.fuzzy=Fuzzy Match
+command.thankwords.check.message.mention=Mention Match
+command.thankwords.check.message.noMatch=No Match
 command.thankwords.sub.check.notConfident=Confidence Score is to low. A reputation will not be given.
-command.thankwords.sub.check.reference=Reference Message: [Jump](%URL%)
-command.thankwords.sub.check.result=Donator: %DONATOR%\nReceiver: %RECEIVER%
-command.thankwords.sub.list=List all pattern
-command.thankwords.sub.list.list=Following Pattern are active:
-command.thankwords.sub.loadDefault=Load default thankswords for a language.
-command.thankwords.sub.loadDefault.added=Added the following thankwords:
-command.thankwords.sub.loadDefault.arg.language=The language to load thankwords for.
-command.thankwords.sub.loadDefault.available=Available Languages:
-command.thankwords.sub.remove=Delete a pattern
-command.thankwords.sub.remove.arg.pattern=The regex pattern to remove.
-command.thankwords.sub.remove.removed=Removed pattern %PATTERN%.
 command.top.message.empty=There are no entries
 command.top.message.monthTitle=Ranking of %GUILD% for the last 30 days
 command.top.message.total=Ranking of %GUILD%
 command.top.message.weekTitle=Ranking of %GUILD% for the last 7 days.
+command.thankwords.check.message.reference=Reference Message: [Jump](%URL%)
+command.thankwords.check.message.result=Donator: %DONATOR%\nReceiver: %RECEIVER%
+command.thankwords.list.description=List all pattern
+command.thankwords.list.message.list=Following Pattern are active:
+command.thankwords.loaddefault.description=Load default thankswords for a language.
+command.thankwords.loaddefault.message.added=Added the following thankwords:
+command.thankwords.loaddefault.language.description=The language to load thankwords for.
+command.thankwords.loaddefault.message.available=Available Languages:
+command.thankwords.remove.description=Delete a pattern
+command.thankwords.remove.pattern.description=The regex pattern to remove.
+command.thankwords.remove.message.removed=Removed pattern %PATTERN%.
 conversation.cancel=cancel
 conversation.canceled=Conversation canceled.
 conversation.exit=exit
