+++ src/main/resources/locale_es_ES.properties
@@ -275,34 +275,34 @@ command.setup.dialog.setupComplete=¡Instalación completa!
 command.setup.dialog.starting=Iniciando la instalación.
 command.setup.dialog.welcome=Bienvenido a la configuración del bot de reputación.
 command.thankwords.description=Configurar Thankwords
-command.thankwords.error.patternNotFound=El patrón no existe.
-command.thankwords.sub.add=Añadir otro patrón. Los patrones se interpretan como Regex. Un patrón no puede contener espacios.
-command.thankwords.sub.add.added=Añadir patrón %REGEX%.
-command.thankwords.sub.add.arg.pattern=El patrón regex a añadir.
-command.thankwords.sub.check=Comprobar si un mensaje da reputación.
-command.thankwords.sub.check.arg.message=El mensaje a comprobar con las thankwords actuales en el canal actual.
-command.thankwords.sub.check.confidence=Puntuación de confianza: %SCORE%.
-command.thankwords.sub.check.match.answer=Concordancia de respuesta
-command.thankwords.sub.check.match.fuzzy=Concordancia difusa
-command.thankwords.sub.check.match.mention=Concordancia con la mención
-command.thankwords.sub.check.match.noMatch=Sin coincidencia
+command.thankwords.remove.message.patternnotnound=El patrón no existe.
+command.thankwords.add.description=Añadir otro patrón. Los patrones se interpretan como Regex. Un patrón no puede contener espacios.
+command.thankwords.add.message.added=Añadir patrón %REGEX%.
+command.thankwords.add.pattern.description=El patrón regex a añadir.
+command.thankwords.check.description=Comprobar si un mensaje da reputación.
+command.thankwords.check.message.description=El mensaje a comprobar con las thankwords actuales en el canal actual.
+command.thankwords.check.message.confidence=Puntuación de confianza: %SCORE%.
+command.thankwords.check.message.answer=Concordancia de respuesta
+command.thankwords.check.message.fuzzy=Concordancia difusa
+command.thankwords.check.message.mention=Concordancia con la mención
+command.thankwords.check.message.noMatch=Sin coincidencia
 command.thankwords.sub.check.notConfident=La puntuación de confianza es demasiado baja. No se dará una reputación.
-command.thankwords.sub.check.reference=Mensaje de referencia: [Jump](%URL%)
-command.thankwords.sub.check.result=Donador: %DONATOR%\nReceptor: %RECEIVER%
-command.thankwords.sub.list=Lista de todos los patrones
-command.thankwords.sub.list.list=Los siguientes patrones están activos:
-command.thankwords.sub.loadDefault=Cargar palabras de agradecimiento por defecto para un idioma.
-command.thankwords.sub.loadDefault.added=Añadir las siguientes palabras de agradecimiento:
-command.thankwords.sub.loadDefault.arg.language=El idioma para el que se deben cargar las palabras de agradecimiento.
-command.thankwords.sub.loadDefault.available=Idiomas disponibles:
-command.thankwords.sub.remove=Eliminar un patrón
-command.thankwords.sub.remove.arg.pattern=El patrón regex a eliminar.
-command.thankwords.sub.remove.removed=Eliminar patrón %PATTERN%.
 command.top.message.empty=No hay entradas
 command.top.message.monthTitle=Clasificación del %GUILD% en los últimos 30 días
 command.top.message.total=Ranking de %GUILD%
 command.top.message.weekTitle=Clasificación del %GUILD% en los últimos 7 días.
 command.top.mode.description=$command.reputation.description.arg.mode$
+command.thankwords.check.message.reference=Mensaje de referencia: [Jump](%URL%)
+command.thankwords.check.message.result=Donador: %DONATOR%\nReceptor: %RECEIVER%
+command.thankwords.list.description=Lista de todos los patrones
+command.thankwords.list.message.list=Los siguientes patrones están activos:
+command.thankwords.loaddefault.description=Cargar palabras de agradecimiento por defecto para un idioma.
+command.thankwords.loaddefault.message.added=Añadir las siguientes palabras de agradecimiento:
+command.thankwords.loaddefault.language.description=El idioma para el que se deben cargar las palabras de agradecimiento.
+command.thankwords.loaddefault.message.available=Idiomas disponibles:
+command.thankwords.remove.description=Eliminar un patrón
+command.thankwords.remove.pattern.description=El patrón regex a eliminar.
+command.thankwords.remove.message.removed=Eliminar patrón %PATTERN%.
 conversation.cancel=Cancelar
 conversation.canceled=Conversación cancelada.
 conversation.exit=salir
