+++ src/main/resources/locale_fr_FR.properties
@@ -275,34 +275,34 @@ command.setup.dialog.setupComplete=Installation terminée !
 command.setup.dialog.starting=Démarre l'installation.
 command.setup.dialog.welcome=Bienvenue dans la configuration du robot de réputation.
 command.thankwords.description=Configurer Thankwords
-command.thankwords.error.patternNotFound=Le modèle n'existe pas.
-command.thankwords.sub.add=Ajoute un motif. Ceux-ci sont interprétés comme des regex. Pas d'espaces !
-command.thankwords.sub.add.added=Ajouté le motif %REGEX%.
-command.thankwords.sub.add.arg.pattern=Le motif regex à ajouter.
-command.thankwords.sub.check=Vérifie si un message donne une réputation.
-command.thankwords.sub.check.arg.message=Le message à vérifier avec les mots de remerciement actuels dans le canal actuel.
-command.thankwords.sub.check.confidence=Score de confiance : %SCORE%.
-command.thankwords.sub.check.match.answer=Mise en correspondance des réponses
-command.thankwords.sub.check.match.fuzzy=Correspondance floue
-command.thankwords.sub.check.match.mention=Mention Match
-command.thankwords.sub.check.match.noMatch=Pas de match
+command.thankwords.remove.message.patternnotnound=Le modèle n'existe pas.
+command.thankwords.add.description=Ajoute un motif. Ceux-ci sont interprétés comme des regex. Pas d'espaces !
+command.thankwords.add.message.added=Ajouté le motif %REGEX%.
+command.thankwords.add.pattern.description=Le motif regex à ajouter.
+command.thankwords.check.description=Vérifie si un message donne une réputation.
+command.thankwords.check.message.description=Le message à vérifier avec les mots de remerciement actuels dans le canal actuel.
+command.thankwords.check.message.confidence=Score de confiance : %SCORE%.
+command.thankwords.check.message.answer=Mise en correspondance des réponses
+command.thankwords.check.message.fuzzy=Correspondance floue
+command.thankwords.check.message.mention=Mention Match
+command.thankwords.check.message.noMatch=Pas de match
 command.thankwords.sub.check.notConfident=Le score de confiance est trop faible. Une réputation ne sera pas donnée.
-command.thankwords.sub.check.reference=Message de référence : [Sauter](%URL%)
-command.thankwords.sub.check.result=Donateur : %DONATOR%\nReceveur : %RECEIVER%.
-command.thankwords.sub.list=Liste de tous les motifs
-command.thankwords.sub.list.list=Les motifs suivants sont actifs :
-command.thankwords.sub.loadDefault=Charge les mots de remerciement par défaut pour une langue.
-command.thankwords.sub.loadDefault.added=Ajoute les mots de remerciement suivants :
-command.thankwords.sub.loadDefault.arg.language=La langue pour laquelle charger les mots de remerciement.
-command.thankwords.sub.loadDefault.available=Les langues disponibles :
-command.thankwords.sub.remove=Supprimer un motif
-command.thankwords.sub.remove.arg.pattern=Le modèle de regex à supprimer.
-command.thankwords.sub.remove.removed=Modèle supprimé %PATTERN%.
 command.top.message.empty=Il n'y a pas d'entrées
 command.top.message.monthTitle=Classement de %GUILD% pour les 30 derniers jours.
 command.top.message.total=Classement de %GUILD%
 command.top.message.weekTitle=Classement de %GUILD% pour les 7 derniers jours.
 command.top.mode.description=$command.reputation.description.arg.mode$
+command.thankwords.check.message.reference=Message de référence : [Sauter](%URL%)
+command.thankwords.check.message.result=Donateur : %DONATOR%\nReceveur : %RECEIVER%.
+command.thankwords.list.description=Liste de tous les motifs
+command.thankwords.list.message.list=Les motifs suivants sont actifs :
+command.thankwords.loaddefault.description=Charge les mots de remerciement par défaut pour une langue.
+command.thankwords.loaddefault.message.added=Ajoute les mots de remerciement suivants :
+command.thankwords.loaddefault.language.description=La langue pour laquelle charger les mots de remerciement.
+command.thankwords.loaddefault.message.available=Les langues disponibles :
+command.thankwords.remove.description=Supprimer un motif
+command.thankwords.remove.pattern.description=Le modèle de regex à supprimer.
+command.thankwords.remove.message.removed=Modèle supprimé %PATTERN%.
 conversation.cancel=annuler
 conversation.canceled=Conversation annulée.
 conversation.exit=sortir
