+++ src/main/resources/locale_pt_PT.properties
@@ -275,34 +275,34 @@ command.setup.dialog.setupComplete=Configuração completa!
 command.setup.dialog.starting=Arrancar setup.
 command.setup.dialog.welcome=Bem-vindos à configuração do bot de reputação.
 command.thankwords.description=Configurar palavras.de.agradecimento
-command.thankwords.error.patternNotFound=Pattern não existe.
-command.thankwords.sub.add=Adicionar outro padrão. Padrões são interpretados como Regex. Um padrão não pode conter espaços.
-command.thankwords.sub.add.added=Adicionar padrão %REGEX%.
-command.thankwords.sub.add.arg.pattern=O padrão regex a adicionar.
-command.thankwords.sub.check=Cheque se uma mensagem daria reputação.
-command.thankwords.sub.check.arg.message=A mensagem a verificar com as palavras de agradecimento actuais no canal actual.
-command.thankwords.sub.check.confidence=Nota de Confiança: %SCORE%
-command.thankwords.sub.check.match.answer=Aresposta Match
-command.thankwords.sub.check.match.fuzzy=Fuzzy Match
-command.thankwords.sub.check.match.mention=Menção Match
-command.thankwords.sub.check.match.noMatch=Sem correspondência
+command.thankwords.remove.message.patternnotnound=Pattern não existe.
+command.thankwords.add.description=Adicionar outro padrão. Padrões são interpretados como Regex. Um padrão não pode conter espaços.
+command.thankwords.add.message.added=Adicionar padrão %REGEX%.
+command.thankwords.add.pattern.description=O padrão regex a adicionar.
+command.thankwords.check.description=Cheque se uma mensagem daria reputação.
+command.thankwords.check.message.description=A mensagem a verificar com as palavras de agradecimento actuais no canal actual.
+command.thankwords.check.message.confidence=Nota de Confiança: %SCORE%
+command.thankwords.check.message.answer=Aresposta Match
+command.thankwords.check.message.fuzzy=Fuzzy Match
+command.thankwords.check.message.mention=Menção Match
+command.thankwords.check.message.noMatch=Sem correspondência
 command.thankwords.sub.check.notConfident=A pontuação de confiança é muito baixa. Não será dada uma reputação.
-command.thankwords.sub.check.reference=Mensagem de referência: [Saltar](%URL%)
-command.thankwords.sub.check.result=Donatador: %DONATOR%\nReceptor: %RECEIVER%
-command.thankwords.sub.list=Listar todos os padrões
-command.thankwords.sub.list.list=Os seguintes Padrões estão activos:
-command.thankwords.sub.loadDefault=Carregar palavras de agradecimento por defeito para uma língua.
-command.thankwords.sub.loadDefault.added=Acrescentadas as seguintes palavras de agradecimento:
-command.thankwords.sub.loadDefault.arg.language=A língua para a qual carregar as palavras de agradecimento.
-command.thankwords.sub.loadDefault.available=Línguas disponíveis:
-command.thankwords.sub.remove=Eliminar um padrão
-command.thankwords.sub.remove.arg.pattern=O padrão regex a remover.
-command.thankwords.sub.remove.removed=Padrão removido %PATTERN%.
 command.top.message.empty=Não há entradas
 command.top.message.monthTitle=Ranking de %GUILD% para os últimos 30 dias
 command.top.message.total=Ranking de %GUILD%
 command.top.message.weekTitle=Ranking de %GUILD% para os últimos 7 dias.
 command.top.mode.description=$command.reputation.description.arg.mode$
+command.thankwords.check.message.reference=Mensagem de referência: [Saltar](%URL%)
+command.thankwords.check.message.result=Donatador: %DONATOR%\nReceptor: %RECEIVER%
+command.thankwords.list.description=Listar todos os padrões
+command.thankwords.list.message.list=Os seguintes Padrões estão activos:
+command.thankwords.loaddefault.description=Carregar palavras de agradecimento por defeito para uma língua.
+command.thankwords.loaddefault.message.added=Acrescentadas as seguintes palavras de agradecimento:
+command.thankwords.loaddefault.language.description=A língua para a qual carregar as palavras de agradecimento.
+command.thankwords.loaddefault.message.available=Línguas disponíveis:
+command.thankwords.remove.description=Eliminar um padrão
+command.thankwords.remove.pattern.description=O padrão regex a remover.
+command.thankwords.remove.message.removed=Padrão removido %PATTERN%.
 conversation.cancel=cancelar
 conversation.canceled=A conversa foi cancelada.
 conversation.exit=saída
