+++ src/main/resources/locale_ru_RU.properties
@@ -275,34 +275,34 @@ command.setup.dialog.setupComplete=Настройка завершена!
 command.setup.dialog.starting=Начинаю настройку.
 command.setup.dialog.welcome=Добро пожаловать в настройку "Reputation Bot".
 command.thankwords.description=Настроить слова благодарности
-command.thankwords.error.patternNotFound=Шаблон не существует.
-command.thankwords.sub.add=Добавить новый шаблон. Шаблоны интерпретируются как Regex. Шаблон не может содержать пробелы.
-command.thankwords.sub.add.added=Добавлен шаблон %REGEX%.
-command.thankwords.sub.add.arg.pattern=Шаблон regex для добавления.
-command.thankwords.sub.check=Проверьте, повысит ли сообщение репутацию.
-command.thankwords.sub.check.arg.message=Сообщение, которое нужно проверить с текущими thankwords в текущем канале.
-command.thankwords.sub.check.confidence=Уровень доверия: %SCORE%
-command.thankwords.sub.check.match.answer=Соответствие ответа
-command.thankwords.sub.check.match.fuzzy=Неопределенное соответствие
-command.thankwords.sub.check.match.mention=Соответствие упоминания
-command.thankwords.sub.check.match.noMatch=Нет соответствия
+command.thankwords.remove.message.patternnotnound=Шаблон не существует.
+command.thankwords.add.description=Добавить новый шаблон. Шаблоны интерпретируются как Regex. Шаблон не может содержать пробелы.
+command.thankwords.add.message.added=Добавлен шаблон %REGEX%.
+command.thankwords.add.pattern.description=Шаблон regex для добавления.
+command.thankwords.check.description=Проверьте, повысит ли сообщение репутацию.
+command.thankwords.check.message.description=Сообщение, которое нужно проверить с текущими thankwords в текущем канале.
+command.thankwords.check.message.confidence=Уровень доверия: %SCORE%
+command.thankwords.check.message.answer=Соответствие ответа
+command.thankwords.check.message.fuzzy=Неопределенное соответствие
+command.thankwords.check.message.mention=Соответствие упоминания
+command.thankwords.check.message.noMatch=Нет соответствия
 command.thankwords.sub.check.notConfident=Уровень доверия слишком низкий. Репутация не была повышена.
-command.thankwords.sub.check.reference=Исходное сообщение: [Jump](%URL%)
-command.thankwords.sub.check.result=Благодарящий: %DONATOR%\nПолучатель: %RECEIVER%
-command.thankwords.sub.list=Список всех шаблонов
-command.thankwords.sub.list.list=Следующий шаблон используется:
-command.thankwords.sub.loadDefault=Загрузить стандартные слова благодарности для языка.
-command.thankwords.sub.loadDefault.added=Добавлено новое слово благодарности:
-command.thankwords.sub.loadDefault.arg.language=Язык, для которого следует загружать благодарственные слова.
-command.thankwords.sub.loadDefault.available=Доступные языки:
-command.thankwords.sub.remove=Удалить шаблон
-command.thankwords.sub.remove.arg.pattern=Шаблон regex для удаления.
-command.thankwords.sub.remove.removed=Удален шаблон %PATTERN%.
 command.top.message.empty=Нет записей
 command.top.message.monthTitle=Ранжирование %GUILD% за последние 30 дней
 command.top.message.total=Иерархия %GUILD%
 command.top.message.weekTitle=Ранжирование %GUILD% за последние 7 дней.
 command.top.mode.description=$command.reputation.description.arg.mode$
+command.thankwords.check.message.reference=Исходное сообщение: [Jump](%URL%)
+command.thankwords.check.message.result=Благодарящий: %DONATOR%\nПолучатель: %RECEIVER%
+command.thankwords.list.description=Список всех шаблонов
+command.thankwords.list.message.list=Следующий шаблон используется:
+command.thankwords.loaddefault.description=Загрузить стандартные слова благодарности для языка.
+command.thankwords.loaddefault.message.added=Добавлено новое слово благодарности:
+command.thankwords.loaddefault.language.description=Язык, для которого следует загружать благодарственные слова.
+command.thankwords.loaddefault.message.available=Доступные языки:
+command.thankwords.remove.description=Удалить шаблон
+command.thankwords.remove.pattern.description=Шаблон regex для удаления.
+command.thankwords.remove.message.removed=Удален шаблон %PATTERN%.
 conversation.cancel=отмена
 conversation.canceled=Беседа отменена.
 conversation.exit=выход
