+++ src/test/java/de/chojo/repbot/commands/ThankwordsTest.java
@@ -1,5 +1,6 @@
 package de.chojo.repbot.commands;
 
+import de.chojo.repbot.commands.thankwords.Thankwords;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
